requirements.txt
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-extensions==2.2.7
Pillow==7.0.0
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2==2.8.4
django-cors-headers==3.2.0